build_tools/api_codegen/requirements.txt
jinja2>=3.1.0
pyyaml>=6.0
pytest>=7.0
requests>=2.28.0
urllib3>=1.26.0
